+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 * Added support for camel-case identifiers in filter queries. Identifiers in
   query strings are now consistent with identifiers in Node source code.
   Snake-case names in query strings are deprecated.
+* Changed method `all` to return a standard async iterator.
+  The previous callback-based interface still works
+  but is deprecated.
 
 ## 1.4.0 (20180308)